requirements.txt
# requirements.txt
Flask
flask-jwt-extended
mysql-connector-python
flasgger